Commit allows turning ggtags support off
Currently all decisions about whether GNU global and ggtags-mode should
be used are based on bound predicates. The result is that if ggtags-mode
has ever been enabled in the current emacs session, you are unable to
revert to using etags/ctags. A new variable was added
`projectile-use-gtags' that allows you to explicitly control this
functionality.